# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/question-outputs/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "CreateQuestionOutput",
                "tags": [
                    "questionOutput"
                ],
                "summary": "CreateQuestionOutput",
                "parameters": [
                    {
                        "description": "questionOutput info",
                        "name": "info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/question.CreateQuestionOutputRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "id",
                        "schema": {
                            "$ref": "#/definitions/question.QuestionOutputId"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/question-outputs/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "DeleteQuestionOutput",
                "tags": [
                    "questionOutput"
                ],
                "summary": "DeleteQuestionOutput",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Void",
                        "schema": {
                            "$ref": "#/definitions/question.Void"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/question-outputs/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "GetQuestionOutputById",
                "tags": [
                    "questionOutput"
                ],
                "summary": "GetQuestionOutputById",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "questionOutput info",
                        "schema": {
                            "$ref": "#/definitions/question.GetQuestionOutputResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/question-outputs/{question_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "GetQuestionOutputsByQuestionId",
                "tags": [
                    "questionOutput"
                ],
                "summary": "GetQuestionOutputsByQuestionId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "question_id",
                        "name": "question_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "questionOutput info",
                        "schema": {
                            "$ref": "#/definitions/question.GetAllQuestionOutputsByQuestionIdResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "question.CreateQuestionOutputRequest": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "input_id": {
                    "type": "string"
                },
                "question_id": {
                    "type": "string"
                }
            }
        },
...
        "question.GetAllQuestionOutputsByQuestionIdResponse": {
            "type": "object",
            "properties": {
                "question_outputs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/question.GetQuestionOutputResponse"
                    }
                }
            }
        },
...
        "question.GetQuestionOutputResponse": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "input_id": {
                    "type": "string"
                },
                "question_id": {
                    "type": "string"
                }
            }
        },
...
        "question.QuestionOutputId": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
        "question.Void": {
            "type": "object"
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Gateway",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}